Updates from featuretools v0.26.0 (#1539)
* Change to use GitHub Token rather than GitHub PAT (#1402)

* Update dependency_check.yml

* Update release_notes.rst

* Update dependency_check.yml

* Use builtin secret token with create pull request (#1407)

* Use builtin secret token with create pull request

* Update release_notes.rst

* Use repo scoped token again (#1409)

* Use repo scoped token again

* Update release_notes.rst

* Update latest_dependencies.txt (#1410)


* Lower max depth to 1 if single entity (#1412)

* lower DFS max depth to 1 if single entity

* add test for including seed features with depth greater than max depth

* test max depth=0 doesn't create depth 1 features on single table

* update release notes

* change log entry to user warning and test for warning

* lint

* fix max depth warning in docs

* remove outdated comment

* rework single table assertions to be more readable

* use feature_with_name helper in seed_features test

* lint

* add max_depth=None and max_depth=-1 cases to single table test

* move helper function def out of loop; remove invalid max_depth=None case

* lint

* Drop Python 3.6 support (#1413)

* remove py36 from CI test matrix

* remove warning when importing featuretools about dropping 3.6 support

* remove python 3.6 from setup.py

* remove py36 from list of supported version in installation docs

* remove py36 constraint on dependency

* update release notes

* v0.24.0 (#1414)

* bump version number

* update release notes

* Update latest_dependencies.txt (#1415)


* Separate workflows and unit tests (#1422)

* separate workflows

* update release notes

* fix incorrect word

* update numpy req

* separate link check

* remove dask separation

* copy from main

* release notes

* Add minimum dependency generator GitHub Action (#1428)

* add min deps checker

* update release notes

* fix filename

* generate auto PR

* update latest dep check

* file rename

* better release notes

* move to 1 folder

* fix fastparquet?

* Update minimum dependencies (#1431)


* Bump pyyaml from 3.12 to 5.4 in /featuretools/tests/requirement_files (#1433)

* Bump pyyaml from 3.12 to 5.4 in /featuretools/tests/requirement_files

Bumps [pyyaml](https://github.com/yaml/pyyaml) from 3.12 to 5.4.
- [Release notes](https://github.com/yaml/pyyaml/releases)
- [Changelog](https://github.com/yaml/pyyaml/blob/master/CHANGES)
- [Commits](https://github.com/yaml/pyyaml/compare/3.12...5.4)


* Update requirements.txt

* Update release_notes.rst

* Update release_notes.rst


* Update nbsphinx version to resolve docs build issue (#1436)

* update release note for test

* update release notes

* pin markupsafe version

* update nbsphinx version and remove markupsafe

* update release notes

* Update latest dependencies (#1437)


* Update latest dependencies (#1439)


* Bump psutil requirement (#1438)

* Bump psutil requirement

* Update release_notes.rst

* Update minimum dependencies (#1443)

* Add unit tests against minimum dependencies (#1432)

* Fix numpy installation for minimum unit tests (#1445)

* Update latest dependencies (#1446)


* Update latest dependencies (#1448)


* v0.24.1 (#1450)

* Update latest dependencies (#1454)


* Update latest dependencies (#1455)

* Bump urllib3 from 1.26.4 to 1.26.5 in /featuretools/tests/requirement_files (#1457)

* Bump urllib3 in /featuretools/tests/requirement_files

Bumps [urllib3](https://github.com/urllib3/urllib3) from 1.26.4 to 1.26.5.
- [Release notes](https://github.com/urllib3/urllib3/releases)
- [Changelog](https://github.com/urllib3/urllib3/blob/main/CHANGES.rst)
- [Commits](https://github.com/urllib3/urllib3/compare/1.26.4...1.26.5)

---
updated-dependencies:
- dependency-name: urllib3
  dependency-type: direct:production
...


* Update test-requirements.txt

* Update release_notes.rst


* Update alteryx_open_src_update_checker to 2.0.0 (#1460)

* Update setup.py

* Update __init__.py

* Update release_notes.rst

* Update setup.py

* Update install_test.yml

* double for loop

* Update latest dependencies (#1464)


* Add get_valid_primitives function (#1462)

* add function skeleton

* add tests

* add get_valid_primitives and update tests

* add test and fix typo

* update release notes

* add test for non-str invalid primitive

* remove unused code from custom primitives

* lint

* remove unused var names and avoid erroring due to compatibility

* rework compatibility check

* make ft.get_valid_primitives callable, add to API reference, add note to docstring

* make get_entityset_type private

* Bump minimum pip from 19.0.2 to 21.1.2 (#1475)

* Bump pip from 19.0.2 to 19.2 in /featuretools/tests/requirement_files

Bumps [pip](https://github.com/pypa/pip) from 19.0.2 to 19.2.
- [Release notes](https://github.com/pypa/pip/releases)
- [Changelog](https://github.com/pypa/pip/blob/main/NEWS.rst)
- [Commits](https://github.com/pypa/pip/compare/19.0.2...19.2)

---
updated-dependencies:
- dependency-name: pip
  dependency-type: direct:production
...


* Update test-requirements.txt

* Update test-requirements.txt

* Update minimum_test_requirements.txt

* Update release_notes.rst


* Add dataframe_type property to EntitySet (#1473)

* add dataframe_type property

* remove _get_entityset_type

* update if not pandas entityset checks in tests

* add docstring to dataframe_type

* update release notes

* rework dataframe_type logic

* add test cases

* use dataframe_type in more tests

* remove some unused ks imports

* more test updates

* fix faulty comparison in tests

* v0.25.0 (#1485)

* bump version number

* update release notes

* Update latest dependencies (#1487)

* Update latest dependencies (#1499)


* Fix docs to avoid logging demos (#1498)

* set testing header to prevent logging

* add library to url

* release notes

* release notes


* Update latest dependencies (#1500)

* Update latest dependencies (#1502)


* Update latest dependencies (#1503)


* Add replace_inf_values util function (#1505)

* add replace_inf_values util function

* update release notes

* fix release notes

* add optional columns parameter to function

* lint fix

* Test compatibility with upcoming pandas release 1.3.0 (#1492)

* update requirements

* comment at local error

* fix test_transform error

* fix boolean conversion error

* remove requirements change

* fix timezone warning

* fix astype warning and use view

* Add release note

* Update latest dependencies (#1520)


* Add URL and Email Address primitives (#1508)

* Update latest dependencies (#1524)


* Primitive options include entities overrides exclude entities (#1518)

* update ignore_entity for primitive

* update variable_filter to return True if entity in include_entities

* update release notes

* Update TLD list and add license for email file (#1531)

* add license to primitive data

* update TLD list

* update release notes

* typo

* update TLD list

* v0.26.0 (#1525)

* bump version

* update release notes

* make underline longer

* alphabetize contributors

* Update docs/source/release_notes.rst

* Update docs/source/release_notes.rst



* Update latest dependencies (#1534)

* uncomment future release

* replace target_entity in a few tests

* delete test_entity.py again

* fix include_over_exclude test

* put Fixes section back in the changelog

+++ featuretools/tests/computational_backend/test_calculate_feature_matrix.py
@@ -1612,7 +1612,7 @@ def test_some_instances_not_in_data(pd_es):
 
     assert all(fm.index.values == cutoff_time["instance_id"].values)
     for x, y in zip(fm.columns, [ifeat_answer, prop_answer, dfeat_answer]):
-        np.testing.assert_array_equal(fm[x], y)
+        np.testing.assert_array_equal(fm[x].astype('float64'), y)
 
 
 def test_missing_instances_with_categorical_index(pd_es):
